Renderer: Move vue-i18n to its own file
